+++ common/src/main/java/io/github/mortuusars/exposure/network/Packets.java
@@ -5,6 +5,8 @@ import com.google.common.base.Preconditions;
 import dev.architectury.injectables.annotations.ExpectPlatform;
 import io.github.mortuusars.exposure.network.packet.IPacket;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.server.players.PlayerList;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Predicate;
 
@@ -17,6 +19,13 @@ public class Packets {
     @ExpectPlatform
     public static void sendToClient(IPacket packet, ServerPlayer player) {
         throw new AssertionError();
+    }
+
+    public static void sendToClients(IPacket packet, PlayerList playerList, @Nullable ServerPlayer excludedPlayer) {
+        for (ServerPlayer player : playerList.getPlayers()) {
+            if (player != excludedPlayer)
+                sendToClient(packet, player);
+        }
     }
 
     public static void sendToClients(IPacket packet, ServerPlayer origin, Predicate<ServerPlayer> filter) {